Fixing a retain cycle related ZincEvents
An event held a strong reference to it's source. Unfortunately, the
source would often hold a list of it's events, creating a retain cycle.
The source was just used for informational purproses, so now I'm
converting it to a string (via `description`) and storing it as a
regular event attribute.