Error while Cloning fix
If a bundle failed to clone (for whatever reason) it's status was left at
'cloning'. This would cause Zinc to try to resume cloning again later. However,
it would keep attempting to clone the exact bundle *version*. This can be a
problem if that version was deleted. Now, if cloning fails, it sets the bundle
status back to 'None', and the regular tracking logic will pick up a new version.